pyproject.toml
[tool.poetry]
name = "evergreen.py"
version = "3.0.0"
description = "Python client for the Evergreen API"
authors = [
    "Alexander Costas <alexander.costas@mongodb.com>",
    "Cristopher Stauffer <cristopher.stauffer@mongodb.com>",
    "David Bradford <david.bradford@mongodb.com>",
    "Jeff Zambory <jeff.zambory@mongodb.com>",
    "Jim O'Leary <jim.oleary@mongodb.com>",
    "Lydia Stepanek <lydia.stepanek@mongodb.com>",
]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/evergreen-ci/evergreen.py"
packages = [
    { include = "evergreen", from = "src" },
]

[tool.poetry.scripts]
evg-api = "evergreen.cli.main:main"

[tool.poetry.dependencies]
python = "^3.6.1"
Click = "^7"
python-dateutil = "^2"
PyYAML = "^5"
requests = "^2"
structlog = ">=19"
tenacity = ">=5"
pydantic = ">=1.7"

[tool.poetry.dev-dependencies]
pytest = "6.0.2"
pytest-black = "^0.3.7"
pytest-cov = "^2.8"
pytest-flake8 = "^1.0"
pytest-mypy = "^0.8"
pytest-pydocstyle = "^2.0"
sphinx = "^3.1.1"
sphinx-readable-theme = "^1.3.0"
ghpusher = "^0.2.2"
black = "^20.8b1"
isort = "^5.5.4"
pytest-isort = "^1.2.0"
flake8-bugbear = "^20.1.4"
pre-commit = "^2.7.1"

[tool.pytest.ini_options]
flake8-ignore = "W605 W503 W291 E203 E501 F821"
addopts = "--flake8 --black --pydocstyle --mypy --isort"
testpaths = [
    "src",
    "tests",
]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
